Claud

+++ back-end/local-app/src/main/java/br/com/novaalianca/mnss/localapp/domain/kds/KdsTicketItemResponse.java
@@ -1,11 +1,12 @@
 package br.com.novaalianca.mnss.localapp.domain.kds;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record KdsTicketItemResponse(
     UUID id,
     String productName,
-    Double quantity,
+    BigDecimal quantity,
     String observation,
     KdsTicketStatus status
 ) {}
